selenium/requirements.txt
selenium
python-dotenv
pymongo
pymongo[srv]
requests
pyvirtualdisplay